// src/app/models/character.ts
export interface Character {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
  // Propiedades adicionales para los componentes Material
  powerLevel?: number;
  experiencePoints?: number;
}